got the refactor back, tabs includes stacks now. Added the UI of SearchScreen
