+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/CustomBackButton/constants/constants.ts
