+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/CustomBackButton/styles.ts
