+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/CustomHeader/index.tsx
