+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/CustomHeader/styles.ts
