+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/CustomSearchButton/constants/constants.ts
