+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/CustomSearchButton/index.tsx
