+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/CustomSearchButton/styles.ts
