+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/SearchBar/constants/constants.ts
