+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/SearchBar/index.tsx
