+++ src/app/components/AppNavigator/components/LibraryStackScreen/components/SearchBar/styles.ts
