+++ src/app/components/AppNavigator/components/OtherScreen/index.tsx
