+++ src/app/components/AppNavigator/components/OtherScreen/styles.ts
