+++ src/app/components/AppNavigator/components/TabbarIcon/index.tsx
