+++ src/app/components/AppNavigator/components/TabbarIcon/styles.ts
