+++ src/app/components/AppNavigator/constants/constants.ts
@@ -1,10 +1,12 @@
-import CustomHeader from "../components/LibraryTabScreen/components/CustomHeader";
-import { Platform } from "react-native";
-import { white } from "src/constants/colors";
-import { StackNavigationOptions } from "@react-navigation/stack";
+import Routes from 'src/constants/routes';
+import OtherScreen from '../components/OtherScreen';
+import LibraryStackScreen from '../components/LibraryStackScreen';
 
-export const BASE_OPTIONS = {
-  headerBackground: CustomHeader,
-  headerTintColor: white,
-  headerTitleAlign: Platform.OS === 'ios' ? 'center' : 'left',
-} as StackNavigationOptions;
+export const TABS = [Routes.Library, Routes.Wishlist, Routes.AddNew, Routes.Rentals, Routes.Settings];
+export const SCREENS = {
+  [Routes.Library]: LibraryStackScreen,
+  [Routes.Wishlist]: OtherScreen,
+  [Routes.AddNew]: OtherScreen,
+  [Routes.Rentals]: OtherScreen,
+  [Routes.Settings]: OtherScreen
+};
